BigQuery example for Entur
- Documented BigQuery example for Entur
- Update documentation and dependencies
