+++ README.md
@@ -34,3 +34,29 @@ poetry run pre-commit run --all-files
 Hvis noen av pre-commit sjekkene feiler, så prøv å kjøre pre-commit kommandoen på nytt.
 Den fikser som regel det meste, men må kjøres på nytt for å gå feilfritt gjennom med de
 korrigerte filene.
+
+### Jupyter Notebooks
+
+The files ending with `_ipynb.py` are jupyter notebooks
+stored as plain python files, using `jupytext`. To open them as Jupyter notebooks,
+right-click on them in JupyterLab and select Open With &rarr; Notebook.
+
+When testing locally, start JupyterLab with this command:
+
+```shell
+poetry run jupter lab
+```
+
+#### Converting files between .ipynb and .py
+
+```shell
+# convert notebook.ipynb to a .py file file in the double percent format
+poetry run jupytext --to py:percent notebook.ipynb
+
+# convert notebook.ipynb to a .py file in the double percent format
+poetry run jupytext --to notebook notebook.ipynb
+```
+
+For VS Code there are extensions for opening a python script as Jupyter Notebook,
+for example:
+[Jupytext for Notebooks](https://marketplace.visualstudio.com/items?itemName=donjayamanne.vscode-jupytext).
